Update composition.md
There isn't a data type called Number anymore, as far as I can tell. If I'm wrong, feel free to disregard. But for example when capturing Revenue (a non-integer), it seems that people are choosing Double as the data type. Number isn't anywhere on the list when building a field in a mixin.